MI5-69 Kreiran UsersService
.Metoda za registraciju prebačena iz AuthenticationServicea u UsersService jer registracija ne spada u autentikaciju, već u upravljanje korisnicima.
