+++ ws/src/main/java/hr/foi/air/szokpt/ws/network/UsersService.kt
@@ -1,0 +1,17 @@
+package hr.foi.air.szokpt.ws.network
+
+import hr.foi.air.core.network.models.SuccessfulResponseBody
+import hr.foi.air.core.register.RegistrationBody
+import hr.foi.air.szokpt.ws.models.RegistrationResponse
+import retrofit2.Call
+import retrofit2.http.Body
+import retrofit2.http.Header
+import retrofit2.http.POST
+
+interface UsersService {
+    @POST("/register")
+    fun register(
+        @Header("Authorization") authHeader: String,
+        @Body registrationBody: RegistrationBody
+    ): Call<SuccessfulResponseBody<RegistrationResponse>>
+}
